Decomposited global imports, fixed warnings, updates
