+++ package.json
@@ -1,5 +1,6 @@
 {
   "name": "wavelovers",
+  "description": "Wavelovers. Use your device vibration correctly. Make a massager out of a gamepad.",
   "version": "1.0.0",
   "author": "Eugene Serb",
   "license": "GNU GPL v3",
