+++ src/models/IGamepad.ts
@@ -1,0 +1,10 @@
+import { IVibrationActuator } from '@/models/IVibrationActuator';
+
+export interface IGamepad {
+    id: string;
+    index: number;
+    timestamp: number;
+    connected: boolean;
+    vibrationActuator: IVibrationActuator;
+}
+
