+++ src/models/IGamepadEvent.ts
@@ -1,0 +1,6 @@
+import { IGamepad } from '@/models/IGamepad';
+
+export interface IGamepadEvent {
+    gamepad: IGamepad;
+}
+
