+++ src/models/IVibrationActuator.ts
@@ -1,0 +1,8 @@
+import { TPatternUnit } from '@/models/TPatternUnit';
+
+export interface IVibrationActuator {
+    type: string;
+    reset(): void;
+    playEffect(mode: string, pattern: TPatternUnit): void;
+}
+
