+++ src/models/TPattern.ts
@@ -1,0 +1,9 @@
+import { TPatternUnit } from '@/models/TPatternUnit';
+
+export type TPattern = {
+    name: string;
+    type: string;
+    icon: string;
+    pattern: TPatternUnit[];
+}
+
