+++ src/models/TPatternUnit.ts
@@ -1,0 +1,7 @@
+export type TPatternUnit = {
+    startDelay: number,
+    duration: number,
+    weakMagnitude: number,
+    strongMagnitude: number,
+}
+
